MSFT: explicit cast

+++ ArmPkg/Library/ArmMmuLib/Arm/ArmMmuLibCore.c
@@ -231,7 +231,7 @@ FillTranslationTable (
     return;
   }
 
-  PhysicalBase = MemoryRegion->PhysicalBase;
+  PhysicalBase = (UINT32) MemoryRegion->PhysicalBase;
   RemainLength = MIN(MemoryRegion->Length, SIZE_4GB - PhysicalBase);
 
   switch (MemoryRegion->Attributes) {
@@ -288,7 +288,7 @@ FillTranslationTable (
       PhysicalBase += TT_DESCRIPTOR_SECTION_SIZE;
       RemainLength -= TT_DESCRIPTOR_SECTION_SIZE;
     } else {
-      PageMapLength = MIN (RemainLength, TT_DESCRIPTOR_SECTION_SIZE -
+      PageMapLength = (UINT32) MIN (RemainLength, TT_DESCRIPTOR_SECTION_SIZE -
                                          (PhysicalBase % TT_DESCRIPTOR_SECTION_SIZE));
 
       // Case: Physical address aligned on the Section Size (1MB) && the length
@@ -549,7 +549,7 @@ UpdatePageEntries (
   FirstLevelTable = (ARM_FIRST_LEVEL_DESCRIPTOR *)ArmGetTTBR0BaseAddress ();
 
   // Calculate number of 4KB page table entries to change
-  NumPageEntries = Length / TT_DESCRIPTOR_PAGE_SIZE;
+  NumPageEntries = (UINT32) Length / TT_DESCRIPTOR_PAGE_SIZE;
 
   // Iterate for the number of 4KB pages to change
   Offset = 0;
@@ -685,7 +685,7 @@ UpdateSectionEntries (
   ASSERT (FirstLevelIdx < TRANSLATION_TABLE_SECTION_COUNT);
 
   // calculate number of 1MB first level entries this applies to
-  NumSections = Length / TT_DESCRIPTOR_SECTION_SIZE;
+  NumSections = (UINT32) Length / TT_DESCRIPTOR_SECTION_SIZE;
 
   // iterate through each descriptor
   for(i=0; i<NumSections; i++) {
